requirements.txt
ipykernel==6.29.5

# -------------------
# LangChain / LangGraph
# -------------------
langgraph==0.2.34
langchain==0.3.7
langchain-core==0.3.19
langchain-community==0.3.7
langchain-chroma==0.1.4
langchain-groq==0.2.0
langchain-huggingface==0.1.0

# -------------------
# LLM / AI
# -------------------
groq==0.11.0

# -------------------
# Embeddings / Vector DB
# -------------------
chromadb==0.4.24
sentence-transformers==2.7.0
torch==2.2.2

# -------------------
# PDF
# -------------------
PyMuPDF==1.24.9

# -------------------
# Backend
# -------------------
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9

# -------------------
# Database
# -------------------
motor==3.6.0
pymongo==4.10.1

# -------------------
# Email
# -------------------
resend==2.3.0

# -------------------
# Utils
# -------------------
python-dotenv==1.0.1
httpx==0.27.0
numpy==1.26.4

# -------------------
# Audio (optional)
# -------------------
edge-tts==6.1.12
sounddevice==0.4.7
scipy==1.13.1
pygame==2.6.1
